Module sediment::format

Structs

BasinAndStratum

BasinId

ChecksumWriter

GrainAllocationInfo

GrainId

GrainIndex

IndexFile

The root file of a Sediment datastore.

IndexHeader

A header inside of an “Index” file.

LocalGrainId

Stored

StratumFileHeader

The complete header of a Stratum file, which contains the data written to the Sediment database.

StratumHeader

Each Stratum header is 16kb, and describes the state of allocation of each grain within the Stratum.

StratumId

TransactionId

Enums

GrainAllocationStatus

GrainIdError

Traits

ByteUtil